improved naming of object in Helper and created base AdjustSpeedAsTravelHelper
This code just need to be tested, to make sure drive straight still works and then use this to get the swing turn and turn tested.  then can put arc turn in (and Debug) and then AdjustSpeedAsTraveStartSpeedEndSpeed (and Debug) then can create DriveStraightStartEndSpeed and DriveArchStartEndSpeed..
